+++ src/features/Guild/components/Members/index.js
@@ -5,15 +5,15 @@ import type { Guild as GuildType } from 'flowTypes';
 import { Component } from 'react';
 import { connect } from 'react-redux';
 import get from 'lodash/get';
-import times from 'lodash/times';
 
 import UserContentCard from 'common/components/ContentCard/User';
 import PaginatorGrid from 'common/layouts/PaginatorGrid';
 import { selector } from 'features/Guild/guilds.reducer';
 import { fetchGuildMembers } from 'features/Guild/actions';
+import { makeStubItems } from 'lib/paginator';
 
 const MEMBERS_PER_PAGE = 50;
-const STUB_MEMBERS = { rows: times(MEMBERS_PER_PAGE, () => undefined), count: 9999 };
+const STUB_MEMBERS = makeStubItems(12);
 
 const makeKey = (content, index) => (content ? content.accountName : index);