+++ src/features/User/index.js
@@ -11,9 +11,9 @@ import isObject from 'lodash/isObject';
 import filter from 'lodash/filter';
 import T from 'i18n-react';
 import { Link } from 'react-router';
-import times from 'lodash/times';
 import startCase from 'lodash/startCase';
 
+import { makeStubItems } from 'lib/paginator';
 import PaginatorGrid from 'common/layouts/PaginatorGrid';
 import TooltipTrigger from 'common/components/TooltipTrigger';
 import Button from 'common/components/Button';
@@ -26,7 +26,7 @@ import PvpGame from './components/PvpGame';
 import Overview from './components/Overview';
 import Characters from './components/Characters';
 
-const STUB_GUILDS = times(4, () => undefined);
+const STUB_GUILDS = makeStubItems(4);
 
 import {
   fetchUser,
@@ -185,7 +185,7 @@ export default class User extends Component {
           content: (
             <PaginatorGrid
               key="guilds"
-              rows={guilds || STUB_GUILDS}
+              rows={guilds || STUB_GUILDS.rows}
               limit={0}
               count={0}
               action={() => Promise.resolve()}