+++ src/index.js
@@ -4,6 +4,7 @@ import 'normalize.css';
 import ReactDOM from 'react-dom';
 
 import * as ls from 'lib/localStorage';
+import Tooltip from 'common/components/Tooltip';
 
 ls.reset();
 
@@ -16,3 +17,21 @@ ReactDOM.render(
   </Base>,
   document.getElementById('root')
 );
+
+function bootstrapTooltip () {
+  const tooltipContainer = document.createElement('div');
+  if (!document.body) {
+    throw new Error('Document body not loaded!');
+  }
+
+  document.body.appendChild(tooltipContainer);
+
+  ReactDOM.render(
+    <Base>
+      <Tooltip />
+    </Base>,
+    tooltipContainer
+  );
+}
+
+bootstrapTooltip();